requirements.txt
gitdb==0.6.4
GitPython==2.0.2
pkg-resources==0.0.0
smmap==0.9.0
